all: fixup: comparison of integer expressions of different signedness

+++ src/inet/networklayer/configurator/ipv4/Ipv4NetworkConfigurator.cc
@@ -239,7 +239,7 @@ void Ipv4NetworkConfigurator::configureRoutingTable(Node *node)
     auto nodePath = node->getModule()->getFullPath();
     EV_DETAIL << "Configuring routing table" << EV_FIELD(nodePath) << endl;
     EV_DETAIL << "Removing extra routes from routing table" << EV_FIELD(nodePath) << endl;
-    for (size_t i = 0; i < routingTable->getNumRoutes();) {
+    for (int i = 0; i < routingTable->getNumRoutes();) {
         auto route = check_and_cast<Ipv4Route *>(routingTable->getRoute(i));
         if (route->getSourceType() == IRoute::MANUAL && route->getSource() == this) {
             auto predicate = [&] (const Ipv4Route *other) { return equalRoutes(route, other); };
@@ -257,7 +257,7 @@ void Ipv4NetworkConfigurator::configureRoutingTable(Node *node)
             i++;
     }
     EV_DETAIL << "Removing all routes from multicast routing table" << EV_FIELD(nodePath) << endl;
-    for (size_t i = 0; i < routingTable->getNumMulticastRoutes();) {
+    for (int i = 0; i < routingTable->getNumMulticastRoutes();) {
         auto route = routingTable->getMulticastRoute(i);
         if (route->getSourceType() == IMulticastRoute::MANUAL && route->getSource() == this) {
             EV_DETAIL << "Removing multicast route" << EV_FIELD(route) << EV_FIELD(nodePath) << endl;
@@ -293,7 +293,7 @@ void Ipv4NetworkConfigurator::configureRoutingTable(Node *node)
                 used = true;
         if (used) {
             Ipv4MulticastRoute *route = new Ipv4MulticastRoute(*original);
-            for (int j = 0; j < route->getNumOutInterfaces();) {
+            for (unsigned int j = 0; j < route->getNumOutInterfaces();) {
                 if (!contains(node->routingTableNetworkInterfaces, route->getOutInterface(j)->getInterface()))
                     route->removeOutInterface(j);
                 else
@@ -1972,7 +1972,7 @@ void Ipv4NetworkConfigurator::addStaticMulticastRoutes(Topology& topology, Node 
         }
 
         bool foundOutInterface = false;
-        for (int i = 0; i < route->getNumOutInterfaces(); i++)
+        for (unsigned int i = 0; i < route->getNumOutInterfaces(); i++)
             if (route->getOutInterface(i)->getInterface() == outInterfaceInfo->networkInterface)
                 foundOutInterface = true;
         if (!foundOutInterface) {
